requirements.txt
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
Bottleneck==1.3.8
bs4==0.0.2
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.0
cssselect==1.2.0
curl_cffi==0.11.3
cycler==0.12.1
decorator==5.1.1
fake-useragent==1.5.1
feedparser==6.0.11
financetoolkit==1.8.4
fonttools==4.50.0
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.42.0
lxml==5.1.0
matplotlib==3.8.3
multitasking==0.0.11
mysql-connector-python==9.3.0
numba==0.59.0
numexpr==2.9.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
parse==1.20.1
peewee==3.17.1
pillow==10.2.0
platformdirs==4.3.6
protobuf==6.31.1
py==1.11.0
pycparser==2.22
pyee==11.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyppeteer==0.0.25
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
requests-html==0.10.0
retry==0.9.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.18.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.41
threadpoolctl==3.3.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
w3lib==2.1.2
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==15.0.1
wsproto==1.2.0
xarray==2024.2.0
yahoo-fin==0.8.9.1
yfinance==0.2.63
zipp==3.18.1
